casts valores retorno cnab

+++ src/Cnab/Retorno/Cnab400/Banco/Banrisul.php
@@ -194,34 +194,34 @@ class Banrisul extends AbstractRetorno implements RetornoCnab400
         $d->setCarteira($this->rem(108, 108, $detalhe))
             ->setNossoNumero($this->rem(63, 72, $detalhe))
             ->setNumeroDocumento($this->rem(117, 126, $detalhe))
-            ->setNumeroControle($this->rem(38, 62, $detalhe))
+            ->setNumeroControle((int) $this->rem(38, 62, $detalhe))
             ->setOcorrencia($this->rem(109, 110, $detalhe))
             ->setOcorrenciaDescricao(Arr::get($this->ocorrencias, $d->getOcorrencia(), 'Desconhecida'))
             ->setDataOcorrencia($this->rem(111, 116, $detalhe))
             ->setDataVencimento($this->rem(147, 152, $detalhe))
             ->setDataCredito($this->rem(296, 301, $detalhe))
-            ->setCodigoLiquidacao($this->rem(343, 343, $detalhe))
+            ->setCodigoLiquidacao((int) $this->rem(343, 343, $detalhe))
             ->setLinhaRegistro($this->rem(395, 400, $detalhe));
 
         if ($this->usandoCentavos) {
-            $d->setValor($this->rem(153, 165, $detalhe))
-                ->setValorTarifa($this->rem(176, 188, $detalhe))
-                ->setValorOutrasDespesas($this->rem(189, 201, $detalhe))
-                ->setValorAbatimento($this->rem(228, 240, $detalhe))
-                ->setValorDesconto($this->rem(241, 253, $detalhe))
-                ->setValorRecebido($this->rem(254, 266, $detalhe))
-                ->setValorMora($this->rem(267, 279, $detalhe)) // banrisul usa juros e multa no mesmo campo
-                ->setValorOutrasReceitas($this->rem(280, 292, $detalhe));
+            $d->setValor((int) $this->rem(153, 165, $detalhe))
+                ->setValorTarifa((int) $this->rem(176, 188, $detalhe))
+                ->setValorOutrasDespesas((int) $this->rem(189, 201, $detalhe))
+                ->setValorAbatimento((int) $this->rem(228, 240, $detalhe))
+                ->setValorDesconto((int) $this->rem(241, 253, $detalhe))
+                ->setValorRecebido((int) $this->rem(254, 266, $detalhe))
+                ->setValorMora((int) $this->rem(267, 279, $detalhe)) // banrisul usa juros e multa no mesmo campo
+                ->setValorOutrasReceitas((int) $this->rem(280, 292, $detalhe));
         } else {
-            $d->setValor(Util::nFloat($this->rem(153, 165, $detalhe) / 100, 2, false))
-                ->setValorTarifa(Util::nFloat($this->rem(176, 188, $detalhe) / 100, 2, false))
-                ->setValorOutrasDespesas(Util::nFloat($this->rem(189, 201, $detalhe) / 100, 2, false))
-                ->setValorAbatimento(Util::nFloat($this->rem(228, 240, $detalhe) / 100, 2, false))
-                ->setValorDesconto(Util::nFloat($this->rem(241, 253, $detalhe) / 100, 2, false))
-                ->setValorRecebido(Util::nFloat($this->rem(254, 266, $detalhe) / 100, 2, false))
-                ->setValorMora(Util::nFloat($this->rem(267, 279, $detalhe) / 100, 2, false))
-                ->setValorMora(Util::nFloat($this->rem(267, 279, $detalhe)/100, 2, false)) // banrisul usa juros e multa no mesmo campo
-                ->setValorOutrasReceitas(Util::nFloat($this->rem(280, 292, $detalhe)/100, 2, false));
+            $d->setValor((float) Util::nFloat($this->rem(153, 165, $detalhe) / 100, 2, false))
+                ->setValorTarifa((float) Util::nFloat($this->rem(176, 188, $detalhe) / 100, 2, false))
+                ->setValorOutrasDespesas((float) Util::nFloat($this->rem(189, 201, $detalhe) / 100, 2, false))
+                ->setValorAbatimento((float) Util::nFloat($this->rem(228, 240, $detalhe) / 100, 2, false))
+                ->setValorDesconto((float) Util::nFloat($this->rem(241, 253, $detalhe) / 100, 2, false))
+                ->setValorRecebido((float) Util::nFloat($this->rem(254, 266, $detalhe) / 100, 2, false))
+                ->setValorMora((float) Util::nFloat($this->rem(267, 279, $detalhe) / 100, 2, false))
+                ->setValorMora((float) Util::nFloat($this->rem(267, 279, $detalhe)/100, 2, false)) // banrisul usa juros e multa no mesmo campo
+                ->setValorOutrasReceitas((float) Util::nFloat($this->rem(280, 292, $detalhe)/100, 2, false));
         }
 
         /**
