+++ src/Cnab/Retorno/Cnab400/Banco/Bb.php
@@ -218,7 +218,7 @@ class Bb extends AbstractRetorno implements RetornoCnab400
         $d->setCarteira($this->rem(107, 108, $detalhe))
             ->setNossoNumero($this->rem(64, 80, $detalhe))
             ->setNumeroDocumento($this->rem(117, 126, $detalhe))
-            ->setNumeroControle($this->rem(39, 63, $detalhe))
+            ->setNumeroControle((int) $this->rem(39, 63, $detalhe))
             ->setOcorrencia($this->rem(109, 110, $detalhe))
             ->setOcorrenciaDescricao(Arr::get($this->ocorrencias, $d->getOcorrencia(), 'Desconhecida'))
             ->setDataOcorrencia($this->rem(111, 116, $detalhe))
@@ -227,21 +227,21 @@ class Bb extends AbstractRetorno implements RetornoCnab400
             ->setLinhaRegistro($this->rem(395, 400, $detalhe));
 
         if ($this->usandoCentavos) {
-            $d->setValor($this->rem(153, 165, $detalhe))
-                ->setValorTarifa($this->rem(182, 188, $detalhe))
-                ->setValorIOF($this->rem(215, 227, $detalhe))
-                ->setValorAbatimento($this->rem(228, 240, $detalhe))
-                ->setValorDesconto($this->rem(241, 253, $detalhe))
-                ->setValorRecebido($this->rem(254, 266, $detalhe))
-                ->setValorMora($this->rem(267, 279, $detalhe));
+            $d->setValor((int) $this->rem(153, 165, $detalhe))
+                ->setValorTarifa((int) $this->rem(182, 188, $detalhe))
+                ->setValorIOF((int) $this->rem(215, 227, $detalhe))
+                ->setValorAbatimento((int) $this->rem(228, 240, $detalhe))
+                ->setValorDesconto((int) $this->rem(241, 253, $detalhe))
+                ->setValorRecebido((int) $this->rem(254, 266, $detalhe))
+                ->setValorMora((int) $this->rem(267, 279, $detalhe));
         } else {
-            $d->setValor(Util::nFloat($this->rem(153, 165, $detalhe) / 100, 2, false))
-                ->setValorTarifa(Util::nFloat($this->rem(182, 188, $detalhe) / 100, 2, false))
-                ->setValorIOF(Util::nFloat($this->rem(215, 227, $detalhe) / 100, 2, false))
-                ->setValorAbatimento(Util::nFloat($this->rem(228, 240, $detalhe) / 100, 2, false))
-                ->setValorDesconto(Util::nFloat($this->rem(241, 253, $detalhe) / 100, 2, false))
-                ->setValorRecebido(Util::nFloat($this->rem(254, 266, $detalhe) / 100, 2, false))
-                ->setValorMora(Util::nFloat($this->rem(267, 279, $detalhe) / 100, 2, false));
+            $d->setValor((float) Util::nFloat($this->rem(153, 165, $detalhe) / 100, 2, false))
+                ->setValorTarifa((float) Util::nFloat($this->rem(182, 188, $detalhe) / 100, 2, false))
+                ->setValorIOF((float) Util::nFloat($this->rem(215, 227, $detalhe) / 100, 2, false))
+                ->setValorAbatimento((float) Util::nFloat($this->rem(228, 240, $detalhe) / 100, 2, false))
+                ->setValorDesconto((float) Util::nFloat($this->rem(241, 253, $detalhe) / 100, 2, false))
+                ->setValorRecebido((float) Util::nFloat($this->rem(254, 266, $detalhe) / 100, 2, false))
+                ->setValorMora((float) Util::nFloat($this->rem(267, 279, $detalhe) / 100, 2, false));
         }
 
         if ($d->hasOcorrencia('05', '06', '07', '08', '15')) {
